refact: 호스트 API 로직 도메인끼리 묶기
