--- src/features/event/model/hostCreation.ts
@@ -1,6 +1,0 @@
-export interface HostCreationRequest {
-  profileImageUrl: string;
-  hostChannelName: string;
-  hostEmail: string;
-  channelDescription: string;
-}
