+++ src/features/host/api/host.ts
@@ -1,8 +1,17 @@
 import { axiosClient } from '../../../shared/types/api/http-client';
-import { UpdateHostChannelInfoRequest } from '../model/host';
+import { HostCreationRequest, UpdateHostChannelInfoRequest } from '../model/host';
 
-const updateHostInfo = async (hostChannelId: number, dto: UpdateHostChannelInfoRequest) => {
+export const createHost = async (data: HostCreationRequest) => {
+  const response = await axiosClient.post('/host-channels', data);
+  return response.data;
+};
+
+export const updateHostInfo = async (hostChannelId: number, dto: UpdateHostChannelInfoRequest) => {
   const response = await axiosClient.put(`/host-channels/${hostChannelId}`, dto);
   return response.data;
 };
-export default updateHostInfo;
+
+export const deleteHost = async (hostChannelId: number) => {
+  const response = await axiosClient.delete(`/host-channels/${hostChannelId}`);
+  return response.data;
+};
