+++ src/features/host/model/host.ts
@@ -1,3 +1,10 @@
+export interface HostCreationRequest {
+  profileImageUrl: string;
+  hostChannelName: string;
+  hostEmail: string;
+  channelDescription: string;
+}
+
 export interface UpdateHostChannelInfoRequest {
   profileImageUrl: string;
   hostChannelName: string;
